test: 전체 도메인 Rest Docs Test 코드 작성 (#386)
