+++ server/src/test/java/com/codueon/boostUp/domain/utils/ApiDocumentUtils.java
@@ -1,0 +1,17 @@
+package com.codueon.boostUp.domain.utils;
+
+import org.springframework.restdocs.operation.preprocess.OperationRequestPreprocessor;
+import org.springframework.restdocs.operation.preprocess.OperationResponsePreprocessor;
+
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+
+public interface ApiDocumentUtils {
+    static OperationRequestPreprocessor getRequestPreProcessor() {
+        return preprocessRequest(prettyPrint());
+    }
+
+    static OperationResponsePreprocessor getResponsePreProcessor() {
+        return preprocessResponse(prettyPrint());
+    }
+}
